backend_banana_predictor/requirements.txt
tensorflow
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api
matplotlib
numpy
pymongo
python-dotenv